+++ app/lib/pages/tabs/settings_tab.dart
@@ -11,7 +11,6 @@ import 'package:localsend_app/provider/settings_provider.dart';
 import 'package:localsend_app/provider/version_provider.dart';
 import 'package:localsend_app/theme.dart';
 import 'package:localsend_app/util/device_type_ext.dart';
-import 'package:localsend_app/util/native/autostart_helper.dart';
 import 'package:localsend_app/util/native/pick_directory_path.dart';
 import 'package:localsend_app/util/native/platform_check.dart';
 import 'package:localsend_app/widget/custom_dropdown_button.dart';
@@ -24,9 +23,6 @@ import 'package:localsend_app/widget/responsive_list_view.dart';
 import 'package:refena_flutter/refena_flutter.dart';
 import 'package:routerino/routerino.dart';
 import 'package:url_launcher/url_launcher.dart';
-
-final _isLinux = checkPlatform([TargetPlatform.linux]);
-final _isWindows = checkPlatform([TargetPlatform.windows]);
 
 class SettingsTab extends StatelessWidget {
   const SettingsTab();
@@ -100,62 +96,27 @@ class SettingsTab extends StatelessWidget {
                       },
                     ),
                   ],
-                  // Linux autostart is simpler, so a boolean entry is used
-                  if (_isLinux)
+                  if (checkPlatformIsDesktop()) ...[
                     _BooleanEntry(
                       label: t.settingsTab.general.launchAtStartup,
-                      value: vm.settings.launchAtStartup,
-                      onChanged: (b) async {
-                        late bool result;
-                        if (await isLinuxLaunchAtStartEnabled()) {
-                          result = await initDisableAutoStart(vm.settings);
-                        } else {
-                          result = await initEnableAutoStartAndOpenSettings(vm.settings);
-                        }
-                        if (result) {
-                          await ref.notifier(settingsProvider).setLaunchAtStartup(b);
-                        }
-                      },
+                      value: vm.autoStart,
+                      onChanged: (_) => vm.onToggleAutoStart(context),
                     ),
-                  // Windows requires a manual action, so this settings entry is required
-                  if (_isWindows)
-                    _SettingsEntry(
-                      label: t.settingsTab.general.launchAtStartup,
-                      child: TextButton(
-                        style: TextButton.styleFrom(
-                          backgroundColor: Theme.of(context).inputDecorationTheme.fillColor,
-                          shape: RoundedRectangleBorder(borderRadius: Theme.of(context).inputDecorationTheme.borderRadius),
-                          foregroundColor: Theme.of(context).colorScheme.onSurface,
-                        ),
-                        onPressed: () async {
-                          await initDisableAutoStart(vm.settings);
-                          await initEnableAutoStartAndOpenSettings(vm.settings, _isWindows);
-                        },
-                        child: Padding(
-                          padding: const EdgeInsets.symmetric(vertical: 5),
-                          child: Text(t.general.settings, style: Theme.of(context).textTheme.titleMedium),
-                        ),
-                      ),
-                    ),
-                  if (_isWindows || _isLinux)
                     Visibility(
-                      visible: vm.settings.launchAtStartup || _isWindows,
+                      visible: vm.autoStart,
                       maintainAnimation: true,
                       maintainState: true,
                       child: AnimatedOpacity(
-                        opacity: vm.settings.launchAtStartup || _isWindows ? 1.0 : 0.0,
+                        opacity: vm.autoStart ? 1.0 : 0.0,
                         duration: const Duration(milliseconds: 500),
                         child: _BooleanEntry(
                           label: t.settingsTab.general.launchMinimized,
-                          value: vm.settings.autoStartLaunchMinimized,
-                          onChanged: (b) async {
-                            await initDisableAutoStart(vm.settings);
-                            await ref.notifier(settingsProvider).setAutoStartLaunchMinimized(b);
-                            await initEnableAutoStartAndOpenSettings(vm.settings, _isWindows);
-                          },
+                          value: vm.autoStartLaunchHidden,
+                          onChanged: (_) => vm.onToggleAutoStartLaunchHidden(context),
                         ),
                       ),
                     ),
+                  ],
                 ],
                 _BooleanEntry(
                   label: t.settingsTab.general.animations,